requirements.txt
Flask>=2.3
python-dotenv>=1.0
gunicorn>=20.1.0
python-docx
pymongo>=4.6
dnspython>=2.4
itsdangerous>=2.1
Werkzeug>=2.3
